inbound: Cleanup in preparation for route policies #1781
Before making changes that support inbound HTTP route authorization
policies, this change makes some cosmetic changes that will help
minimize the changes needed for new funcationality.

No funtional changes.
